fix(refactor): adjust import locations after move
