+++ src/app/app-routing.module.ts
@@ -67,7 +67,7 @@ export const routes: Routes = [
     resolve: {
       context: ContextResolver
     },
-    loadChildren: './analyze/analyze.module#AnalyzeModule',
+    loadChildren: './space/analyze/analyze.module#AnalyzeModule',
     data: {
       title: 'Analyze'
     }
@@ -80,7 +80,7 @@ export const routes: Routes = [
       context: ContextResolver,
       featureFlagConfig: ExperimentalFeatureResolver
     },
-    loadChildren: './plan/plan.module#PlanModule',
+    loadChildren: './space/plan/plan.module#PlanModule',
     data: {
       title: 'Plan: Backlog',
       featureName: 'Planner'
@@ -94,7 +94,7 @@ export const routes: Routes = [
       context: ContextResolver,
       featureFlagConfig: ExperimentalFeatureResolver
     },
-    loadChildren: './plan/board/board.module#BoardModule',
+    loadChildren: './space/plan/board/board.module#BoardModule',
     data: {
       title: 'Plan: Board',
       featureName: 'Planner'
@@ -107,7 +107,7 @@ export const routes: Routes = [
     resolve: {
       context: ContextResolver
     },
-    loadChildren: './create/create.module#CreateModule',
+    loadChildren: './space/create/create.module#CreateModule',
     data: {
       title: 'Create'
     }
@@ -119,7 +119,7 @@ export const routes: Routes = [
     resolve: {
       context: ContextResolver
     },
-    loadChildren: './space-settings/space-settings.module#SpaceSettingsModule',
+    loadChildren: './space/space-settings/space-settings.module#SpaceSettingsModule',
     data: {
       title: 'Areas'
     }
