+++ src/app/dashboard-widgets/create-work-item-widget/create-work-item-widget-routing.module.ts
@@ -2,7 +2,7 @@ import { CreateWorkItemOverlayComponent } from './create-work-item-overlay/creat
 import { NgModule }  from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { codebasesAddRoutes } from './../../create/codebases/codebases-add/codebases-add-routing.module';
+import { codebasesAddRoutes } from '../../space/create/codebases/codebases-add/codebases-add-routing.module';
 
 const routes: Routes = [
   {
