+++ src/app/shared/recommender-api.provider.ts
@@ -1,4 +1,4 @@
-import { RECOMMENDER_API_URL } from './../analyze/stack/recommender-api';
+import { RECOMMENDER_API_URL } from '../space/analyze/stack/recommender-api';
 import { ApiLocatorService } from './api-locator.service';
 
 let recommenderApiUrlFactory = (api: ApiLocatorService) => {
