+++ src/app/space-wizard/space-wizard.module.ts
@@ -23,7 +23,7 @@ import { SpaceWizardComponent } from './space-wizard.component';
 import { SelectedItemsPipe } from './pipes/selected-items.pipe';
 import { VisibleItemsPipe } from './pipes/visible-items.pipe';
 import { TrustHtmlPipe, TrustStylePipe } from './pipes/safe-html.pipe';
-import { CodebasesService } from '../create/codebases/services/codebases.service';
+import { CodebasesService } from '../space/create/codebases/services/codebases.service';
 
 @NgModule({
   imports: [
