+++ src/app/space/analyze/analyze-overview/analyze-overview.component.less
@@ -1,4 +1,4 @@
-@import (reference) '../../../assets/stylesheets/shared/osio.less';
+@import (reference) '../../../../assets/stylesheets/shared/osio.less';
 .cards-pf { background-color: @color-pf-black-300; }
 .analyze-overview-wrapper {
   background-color: @color-pf-black-300;
