+++ src/app/space/analyze/analyze-overview/analyze-overview.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, OnDestroy, ViewChild, ViewEncapsulation } from '@angular/core';
 import { IWorkflow } from './models/workflow';
-import { IModalHost } from '../../space-wizard/models/modal-host';
-import { SpaceWizardComponent } from '../../space-wizard/space-wizard.component';
+import { IModalHost } from '../../../space-wizard/models/modal-host';
+import { SpaceWizardComponent } from '../../../space-wizard/space-wizard.component';
 import { Context, Contexts } from 'ngx-fabric8-wit';
 
 import {Subscription } from 'rxjs';
