+++ src/app/space/create/codebases/codebases-add/codebases-add.component.less
@@ -1,4 +1,4 @@
-@import (reference) '../../../../assets/stylesheets/shared/osio.less';
+@import (reference) '../../../../../assets/stylesheets/shared/osio.less';
 .details-pull-right { float: right; }
 .details-border {
   border-top: 1px solid;
