+++ src/app/space/create/codebases/codebases-add/codebases-add.component.ts
@@ -16,7 +16,7 @@ import { Codebase } from '../services/codebase';
 import { CodebasesService } from '../services/codebases.service';
 import { GitHubRepoDetails } from "../services/github";
 import { GitHubService } from '../services/github.service';
-import { removeAction } from '../../../app-routing.module';
+import { removeAction } from '../../../../app-routing.module';
 
 @Component({
   encapsulation: ViewEncapsulation.None,
