+++ src/app/space/create/codebases/codebases-item-actions/codebases-item-actions.component.less
@@ -1,2 +1,2 @@
-@import (reference) '../../../../assets/stylesheets/shared/osio.less';
+@import (reference) '../../../../../assets/stylesheets/shared/osio.less';
 .margin-right-20 { margin-right: 20px; }
