+++ src/app/space/create/codebases/codebases-item-workspaces/codebases-item-workspaces.component.less
@@ -1,4 +1,4 @@
-@import (reference) '../../../../assets/stylesheets/shared/osio.less';
+@import (reference) '../../../../../assets/stylesheets/shared/osio.less';
 .busy-indicator {
   display: inline-flex;
   position: relative;
