+++ src/app/space/create/codebases/codebases-item/codebases-item.component.less
@@ -1,2 +1,2 @@
-@import (reference) '../../../../assets/stylesheets/shared/osio.less';
+@import (reference) '../../../../../assets/stylesheets/shared/osio.less';
 .margin-right-30 { margin-right: 30px; }
