+++ src/app/space/create/codebases/codebases-routing.module.ts
@@ -3,7 +3,7 @@ import { RouterModule, Routes } from '@angular/router';
 
 import { codebasesAddRoutes } from './codebases-add/codebases-add-routing.module';
 import { CodebasesComponent } from './codebases.component';
-import { AuthGuard } from '../../shared/auth-guard.service';
+import { AuthGuard } from '../../../shared/auth-guard.service';
 
 const routes: Routes = [
   {
