+++ src/app/space/create/codebases/codebases-toolbar/codebases-toolbar.component.less
@@ -1,4 +1,4 @@
-@import (reference) '../../../../assets/stylesheets/shared/osio.less';
+@import (reference) '../../../../../assets/stylesheets/shared/osio.less';
 .toolbar-header {
   padding: 0 20px;
 }
