+++ src/app/space/create/codebases/codebases.component.less
@@ -1,4 +1,4 @@
-@import (reference) '../../../assets/stylesheets/shared/osio.less';
+@import (reference) '../../../../assets/stylesheets/shared/osio.less';
 .blank-slate-pf { margin-top: 20px; }
 .codebase-item {
   display: inline-flex;
