+++ src/app/space/create/pipelines/pipelines.component.less
@@ -1,4 +1,4 @@
-@import (reference) '../../../assets/stylesheets/shared/osio.less';
+@import (reference) '../../../../assets/stylesheets/shared/osio.less';
 .openshift-links { font-size: .857em; }
 .toolbar-header {
   padding: 0 20px;
