+++ src/app/space/space-settings/areas/areas.component.less
@@ -1,4 +1,4 @@
-@import (reference) '../../../assets/stylesheets/shared/osio.less';
+@import (reference) '../../../../assets/stylesheets/shared/osio.less';
 .blank-slate-pf { margin-top: 20px; }
 .areas-header {
   margin-top: 50px;
