+++ src/app/space/space-settings/areas/create-area-dialog/create-area-dialog.component.less
@@ -1,4 +1,4 @@
-@import (reference) '../../../../assets/stylesheets/shared/osio.less';
+@import (reference) '../../../../../assets/stylesheets/shared/osio.less';
 .create-dialog { padding: 20px; }
 .create-dialog-header {
   position: absolute;
