+++ src/app/space/space-settings/collaborators/add-collaborators-dialog/add-collaborators-dialog.component.less
@@ -1,4 +1,4 @@
-@import (reference) '../../../../assets/stylesheets/shared/osio.less';
+@import (reference) '../../../../../assets/stylesheets/shared/osio.less';
 .add-dialog { padding: 20px; }
 .add-dialog-header {
   position: absolute;
