+++ src/app/space/space-settings/collaborators/collaborators.component.less
@@ -1,4 +1,4 @@
-@import (reference) '../../../assets/stylesheets/shared/osio.less';
+@import (reference) '../../../../assets/stylesheets/shared/osio.less';
 .blank-slate-pf { margin-top: 20px; }
 .collaborators-header {
   margin-top: 50px;
