+++ src/app/space/space-settings/space-settings-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { ContextCurrentUserGuard } from './../shared/context-current-user-guard.service';
+import { ContextCurrentUserGuard } from './../../shared/context-current-user-guard.service';
 import { SpaceSettingsComponent } from './space-settings.component';
 import { SettingsOverviewComponent } from './settings-overview/settings-overview.component';
 import { AreasComponent } from './areas/areas.component';
